% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frac}
\alias{frac}
\title{Affichage LaTeX d'une fraction}
\usage{
frac(x, style = c("inline", "frac", "sfrac", "tfrac", "decimal"),
  digits = 2, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Une liste de nombre decimaux}

\item{style}{Un choix entre quatre facons de representer la fraction
\code{"inline"}; permet d'ecrire la fraction sous la forme a/b
\code{"frac"}; permet d'ecrire la fraction sous la forme frac\{a\}\{b\}
\code{"sfrac"}; permet d'ecrire la fraction sous la forme sfrac\{a\}\{b\}
\code{"decimal"}; permet d'ecrire la fraction sous la forme decimale 0.34...}

\item{digits}{Permet de choisir le nombre de chiffres significatifs a afficher.}
}
\value{
out Une liste contenant la representation sous forme de fraction LaTeX de \code{x}
}
\description{
\code{frac} affiche le code LaTeX sous forme de fraction de tous les nombres
presents dans \code{x}
}
\details{
Cette fonction est surtout utile pour les autres fonctions de la librairie. Elle est appelee lorsque nous faisons
l'affichage de matrices a l'aide la fonction \code{mat}.
}
\examples{
frac(4)
frac(3/7)
frac(3/7, style = "sfrac")
frac(-3/7, style = "inline")
frac(3/7, digits = 2)
}
