% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{mat}
\alias{mat}
\title{Affichage LaTeX d'une matrice}
\usage{
mat(A, B, style = c("inline", "frac", "sfrac", "decimal"),
  bracket = c("parenthese", "crochet", "determinant"), digits = 2)
}
\arguments{
\item{A}{une matrice}

\item{B}{une matrice}

\item{digits}{Permet de choisir le nombre de chiffres significatifs a afficher.}

\item{style;}{Permet de choisir le style d'affichage des fractions. \code{"inline"}; permet d'ecrire la fraction sous la forme a/b.
\code{"frac"}; permet d'ecrire la fraction sous la forme \frac\{a\}\{b\}. \code{"sfrac"}; permet d'ecrire la fraction sous la
forme \sfrac\{a\}\{b\}. \code{"decimal"}; permet d'ecrire la fraction sous la forme decimale 0.34...}

\item{bracket;}{Permet de choisir la facon d'encadrer les matrices. \code{parenthese}; encadre la matrice de parentheses.
\code{crochet}; encadre la matrice de crochets. \code{determinant}; encadre la matrice de valeurs absolues}
}
\value{
Retourne le code LaTex de la matrice
}
\description{
\code{mat} permet l'affichage d'une matrice en code LaTeX. Nous pouvons choisir
le \code{style} pour afficher sous forme de fractions. Nous pouvons afficher
la matrice avec des parentheses, des crochets ou des determinants avec \code{bracket}.
Si la matrice \code{B} est presente, on affiche la matrice augmentee.
}
\seealso{
\code{\link{frac}}
}
