% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel.R
\name{print_mat_var}
\alias{print_mat_var}
\title{Fonction permettant d'afficher une matrice colonne de variables pour l'affichage d'un SEL}
\usage{
print_mat_var(ncol, variables = c("a", "x", "xi"), bracket = c("parenthese",
  "crochet"))
}
\description{
Fonction permettant d'afficher une matrice colonne de variables pour l'affichage d'un SEL
}
