% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel.R
\name{print_sel}
\alias{print_sel}
\title{Affichage d'un systeme d'equations lineaires (SEL)}
\usage{
print_sel(A, B, sel = TRUE, variables = c("x", "a", "xi"),
  style = c("inline", "frac", "sfrac", "decimal"), digits = 2,
  concise = FALSE, bracket = c("crochet", "parenthese"),
  tol = sqrt(.Machine$double.eps))
}
\description{
Affichage d'un systeme d'equations lineaires (SEL)
}
