% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel.R
\name{sel}
\alias{sel}
\title{Affichage d'un systeme d'equations lineaires (SEL)}
\usage{
sel(A, B, sel = TRUE, variables = c("x", "a", "xi"), style = c("inline",
  "frac", "sfrac", "decimal"), digits = 2, bracket = c("crochet",
  "parenthese"), tol = sqrt(.Machine$double.eps))
}
\description{
Affichage d'un systeme d'equations lineaires (SEL)
}
